django/gunicorn/requirements.txt
Django==5.0.6
gunicorn==22.0.0
middleware-apm>=2.0.0